+++ app/components/forms/search_input_component.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+module Forms
+  class SearchInputComponent < ViewComponent::Base
+    def initialize(label:)
+      @label = label
+    end
+  end
+end