+++ app/components/index_header_export_component.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class IndexHeaderExportComponent < ViewComponent::Base
+  def initialize(path:)
+    @path = path
+  end
+end