+++ lib/generators/bo/templates/namespace_policy.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Bo
   class <%= options[:namespace].camelize %>Policy < ActionPolicy::Base
     def show?