+++ lib/generators/bo/templates/policy.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Bo
   module <%= options[:namespace].camelize %>
     class <%= class_name %>Policy < Bo::<%= options[:namespace].camelize %>Policy