+++ lib/generators/bo/utils/translations.rb
@@ -8,6 +8,8 @@ def create_translations
         local => {
           'bo' => {
             'filters' => find_existing_translation('filters', local),
+            'details' => find_existing_translation('details', local),
+            'confirm_delete' => find_existing_translation('confirm_delete', local),
             'record' => {
             'created' => find_existing_translation('created', local),
             'updated' => find_existing_translation('updated', local),
@@ -20,21 +22,23 @@ def create_translations
 
     yaml_string = File.open locale_file
     data = YAML.load yaml_string
-    data[local]['bo'][file_name.underscore.to_s] = {
+    data[local]['bo'][file_name.underscore] = {
       'one' => find_existing_translation(bo_model.to_s.downcase, local),
       'others' => find_existing_translation(bo_model.to_s.pluralize.downcase, local),
+      'new' => find_existing_translation(nil, local),
       'subtitle' => find_existing_translation("list of #{bo_model.to_s.pluralize.downcase}", local),
-      'attributes' => model_attributes(local)
+      'attributes' => model_attributes(data, local)
+
     }
     output = YAML.dump data
     File.write(locale_file, output)
   end
 end
 
-def model_attributes(local)
-  hash = {}
+def model_attributes(data, local)
+  hash = data.dig(local, 'bo', file_name.underscore, 'attributes') || {}
   model_columns.each do |col|
-    hash[col.to_s] = find_existing_translation(col, local)
+    hash[col.to_s] ||= find_existing_translation(col, local)
   end
   hash
 end