+++ CHANGELOG.md
@@ -1,3 +1,32 @@
+### 0.0.39
+- fix paginations errors
+### 0.0.38
+- fix translations errors
+### 0.0.37
+- fix attachments component
+### 0.0.36
+- add better turbo link
+- fix policy namespace error
+- fix attachment component path
+### 0.0.34
+- Raise error if the model argument doesn't exist
+- Add routes check to prevent duplication
+### 0.0.33
+- add components search form date_fields 
+- better html.erb generated
+- Add some translations for actions
+### 0.0.32
+- add components for search form fields
+### 0.0.19
+- disable turbo for devise reset password
+### 0.0.18
+- disable turbo for devise forms
+### 0.0.17
+- disable turbo for devise session form
+### 0.0.16
+- move policy namespace into parent controller
+- add translations for show view and create view
+- remove some params into permited_params 
 ### 0.0.15
 - add new configurable images for bo sidebar
 add into `tybo.rb`